LayoutTests/platform/mac/http/tests/cookies/same-site/popup-cross-site-expected.txt


FAIL '127.0.0.1' is not same-site with 'localhost', so strict samesite cookies are not sent. assert_equals: implicit-strict expected (undefined) undefined but got (string) "1"
